Refactor: Extract scroll-to-top functionality into a shared component
