+++ src/shared/ui/components/scroll-to-top/index.ts
@@ -1,0 +1,1 @@
+export { ScrollToTop } from './scroll-to-top';
