+++ src/widgets/post-detail/ui/post-detail.tsx
@@ -1,46 +1,16 @@
-import { ArrowUpIcon } from '@chakra-ui/icons';
-import { Box, Text, VStack, HStack, Tag, IconButton } from '@chakra-ui/react';
-import { useState, useEffect } from 'react';
+import { Box, Text, VStack, HStack, Tag } from '@chakra-ui/react';
 
 import { useAdminContext } from '@shared/context/admin-context';
 import { getStatusColor } from '@shared/libs';
+import { ScrollToTop } from '@shared/ui/components/scroll-to-top';
 import { type PostDetailType } from '@widgets/post-detail/api';
 
 type PostDetailProps = {
   data: PostDetailType;
 };
 
-const scrollToTop = () => {
-  window.scrollTo({ top: 0, behavior: 'smooth' });
-};
-
 export const PostDetail = ({ data }: PostDetailProps) => {
   const { adminPage: isAdminPage } = useAdminContext();
-  const [showScrollTop, setShowScrollTop] = useState(false);
-
-  useEffect(() => {
-    const sentinel = document.createElement('div');
-    sentinel.style.height = '1px';
-    sentinel.style.visibility = 'hidden';
-
-    document.querySelector('body')?.appendChild(sentinel);
-
-    const observer = new IntersectionObserver(
-      (entries) => {
-        entries.forEach((entry) => {
-          setShowScrollTop(entry.isIntersecting);
-        });
-      },
-      { threshold: 0.1 },
-    );
-
-    observer.observe(sentinel);
-
-    return () => {
-      observer.disconnect();
-      sentinel.remove();
-    };
-  }, []);
 
   return (
     <VStack
@@ -87,20 +57,7 @@ export const PostDetail = ({ data }: PostDetailProps) => {
           </Tag>
         ))}
       </HStack>
-      <IconButton
-        aria-label="Scroll to top"
-        icon={<ArrowUpIcon />}
-        onClick={scrollToTop}
-        position="fixed"
-        bottom="20px"
-        right="20px"
-        borderRadius="full"
-        colorScheme="gray"
-        size="lg"
-        opacity="0.8"
-        _hover={{ opacity: 1 }}
-        display={showScrollTop ? { base: 'flex', md: 'flex' } : 'none'}
-      />
+      <ScrollToTop />
     </VStack>
   );
 };
